requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
django-widget-tweaks==1.4.12
mysqlclient==2.1.0
PyMySQL==1.0.2
sqlparse==0.4.2
